Provide the ability for the browser to use multiple loggers.
This is easily possible due to core's recently added logging framework;
browser will now instantiate a few different loggers that the user can
enable or disable in preferences, as well as control which channels
they log.

It is now possible, for example, to log all channels to the standard
output in addition to the log window, one, both, with certain channels
send to one logger and others to another, in any combination desirable.